separated backends from DI container. backends are imported only in aquisition service now. could decouple config also, but this is fine for now.
