+++ tests/test_backends_allplatforms.py
@@ -5,7 +5,9 @@ import logging
 
 import pytest
 
-from photobooth.services.backends.containers import BackendsContainer
+from photobooth.services.backends.virtualcamera import VirtualCameraBackend
+from photobooth.services.backends.webcamcv2 import WebcamCv2Backend
+from photobooth.services.backends.webcamcv2 import available_camera_indexes as cv2_avail
 from photobooth.services.config import appconfig
 
 from .backends_utils import get_images
@@ -14,26 +16,12 @@ logger = logging.getLogger(name=None)
 
 
 @pytest.fixture()
-def backends() -> BackendsContainer:
+def backend_cv2() -> WebcamCv2Backend:
     # setup
-    backends_container = BackendsContainer()
-    # deliver
-    yield backends_container
-    backends_container.shutdown_resources()
-
-
-def test_get_images_virtualcamera(backends: BackendsContainer):
-    virtualcamera_backend = backends.virtualcamera_backend()
-
-    """get lores and hires images from backend and assert"""
-    get_images(virtualcamera_backend)
-
-
-def test_get_images_webcamcv2(backends: BackendsContainer):
-    from photobooth.services.backends.webcamcv2 import available_camera_indexes
+    backend = WebcamCv2Backend()
 
     logger.info("probing for available cameras")
-    _availableCameraIndexes = available_camera_indexes()
+    _availableCameraIndexes = cv2_avail()
     if not _availableCameraIndexes:
         pytest.skip("no camera found, skipping test")
 
@@ -41,9 +29,31 @@ def test_get_images_webcamcv2(backends: BackendsContainer):
 
     logger.info(f"available camera indexes: {_availableCameraIndexes}")
     logger.info(f"using first camera index to test: {cameraIndex}")
+
     appconfig.backends.cv2_device_index = cameraIndex
 
-    """get lores and hires images from backend and assert"""
+    # deliver
+    backend.start()
+    yield backend
+    backend.stop()
 
-    webcamcv2_backend = backends.webcamcv2_backend()
-    get_images(webcamcv2_backend)
+
+@pytest.fixture()
+def backend_virtual() -> VirtualCameraBackend:
+    # setup
+    backend = VirtualCameraBackend()
+
+    # deliver
+    backend.start()
+    yield backend
+    backend.stop()
+
+
+def test_get_images_virtualcamera(backend_virtual: WebcamCv2Backend):
+    """get lores and hires images from backend and assert"""
+    get_images(backend_virtual)
+
+
+def test_get_images_webcamcv2(backend_cv2: WebcamCv2Backend):
+    """get lores and hires images from backend and assert"""
+    get_images(backend_cv2)
