+++ tests/test_services_exceptions.py
@@ -8,55 +8,10 @@ from unittest.mock import patch
 import pytest
 
 from photobooth.containers import ApplicationContainer
-from photobooth.services.backends.containers import BackendsContainer
-from photobooth.services.backends.virtualcamera import VirtualCameraBackend
 from photobooth.services.containers import ServicesContainer
 from photobooth.services.informationservice import InformationService
 
 logger = logging.getLogger(name=None)
-
-
-@pytest.fixture()
-def backends() -> BackendsContainer:
-    # setup
-    backends_container = BackendsContainer()
-    # deliver
-    yield backends_container
-
-
-def test_simulated_init_exceptions(backends: BackendsContainer):
-    error_mock = mock.MagicMock()
-    error_mock.side_effect = Exception("mock error")
-
-    with patch.object(VirtualCameraBackend, "__init__", error_mock):
-        try:
-            backends.virtualcamera_backend()
-        except Exception as exc:
-            raise AssertionError(f"'simulated_backend' raised an exception, but it should fail in silence {exc}") from exc
-
-
-def test_simulated_start_exceptions(backends: BackendsContainer):
-    error_mock = mock.MagicMock()
-    error_mock.side_effect = Exception("mock error")
-
-    with patch.object(VirtualCameraBackend, "start", error_mock):
-        try:
-            backends.virtualcamera_backend()
-        except Exception as exc:
-            raise AssertionError(f"'simulated_backend' raised an exception, but it should fail in silence {exc}") from exc
-
-
-def test_simulated_stop_exceptions(backends: BackendsContainer):
-    error_mock = mock.MagicMock()
-    error_mock.side_effect = Exception("mock error")
-
-    with patch.object(VirtualCameraBackend, "stop", error_mock):
-        try:
-            backends.virtualcamera_backend()
-            backends.shutdown_resources()
-
-        except Exception as exc:
-            raise AssertionError(f"'simulated_backend' raised an exception, but it should fail in silence {exc}") from exc
 
 
 @pytest.fixture()
